+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+- Moved the `Upward`,`TokenEvent`, `TokenEventDetails` from the `concordium-base`.
+- Moved the `TokenModuleEvent` from the `concordium-base` as `EncodedTokenModuleEvent`.
+- Moved the `TokenModuleRejectReason` from the `concordium-base` as `EncodedTokenModuleRejectReason`.
+
 ## 9.0.1
 
 - Fixes a bug where using `serde` to convert `BlockItemSummary` would result in data loss with regards to transaction sponsor details.